feat: added game ScreenSwitcher & refactor enums imports
