+++ packages/client/src/entities/GameContext/lib/contextInstance.ts
@@ -1,7 +1,6 @@
 import { createContext } from 'react';
 import { TGameContext } from '../types/GameContextTypes';
-import { EGAME_MODE_VALUES } from 'shared/constants/game';
-import { EGAME_SCREEN_VALUES } from 'shared/constants/game/types/ScreensEnum';
+import { EGAME_MODE_VALUES, EGAME_SCREEN_VALUES } from 'shared/constants/game';
 
 export const GameContextInstance = createContext<TGameContext>({
   gameMode: EGAME_MODE_VALUES.FREE_PLAY,
