+++ packages/client/src/entities/GameContext/types/GameContextTypes.ts
@@ -1,5 +1,4 @@
-import { EGAME_MODE_VALUES } from 'shared/constants/game';
-import { EGAME_SCREEN_VALUES } from 'shared/constants/game/types/ScreensEnum';
+import { EGAME_MODE_VALUES, EGAME_SCREEN_VALUES } from 'shared/constants/game';
 
 export type TGameContext = {
   gameMode: EGAME_MODE_VALUES;
