--- packages/client/src/pages/Game/ui/StartGame.tsx
@@ -1,10 +1,0 @@
-import { GameContentContainer } from 'entities/GameContentContainer';
-import { StartGameForm } from 'widgets/StartGameForm';
-
-export const StartGame = () => {
-  return (
-    <GameContentContainer>
-      <StartGameForm />
-    </GameContentContainer>
-  );
-};
