+++ packages/client/src/shared/constants/game/index.ts
@@ -1,2 +1,3 @@
 export { changeGameModeOptions } from './lib/gameModeOptions';
 export { EGAME_MODE_VALUES } from './types/ModeEnum';
+export { EGAME_SCREEN_VALUES } from './types/ScreensEnum';
